backend/ingestion/docx_extractor.py
"""
docx_extractor.py - Extract text and tables from Word (.docx) documents.

Uses python-docx to walk the full document structure:
  - Paragraphs (including styled headings)
  - Tables (with merged cell handling)
  - Text boxes (via XML fallback)
"""

from __future__ import annotations

import logging
from dataclasses import dataclass, field
from pathlib import Path
from typing import Any

from docx import Document
from docx.oxml.ns import qn

logger = logging.getLogger(__name__)


@dataclass
class DocxContent:
    """Structured content extracted from a Word document."""

    paragraphs: list[dict[str, str]] = field(default_factory=list)
    # Each paragraph: {"style": "Heading 1", "text": "...", "level": 1}

    tables: list[list[list[str]]] = field(default_factory=list)
    # List of tables; each table is a list of rows; each row is a list of cell strings

    full_text: str = ""
    # Unified plain-text representation for LLM consumption


class DOCXExtractor:
    """
    Extract structured content from .docx files.

    Preserves heading hierarchy and table structure for downstream
    LIMS mapping.
    """

    # ── Public API ────────────────────────────────────────────────────────────

    def extract(self, file_path: str | Path) -> DocxContent:
        file_path = Path(file_path)
        if not file_path.exists():
            raise FileNotFoundError(f"DOCX file not found: {file_path}")

        doc = Document(str(file_path))
        logger.info("Opened DOCX %s", file_path.name)

        paragraphs = self._extract_paragraphs(doc)
        tables = self._extract_tables(doc)
        full_text = self._build_full_text(paragraphs, tables)

        return DocxContent(paragraphs=paragraphs, tables=tables, full_text=full_text)

    def extract_full_text(self, file_path: str | Path) -> str:
        return self.extract(file_path).full_text

    # ── Private helpers ───────────────────────────────────────────────────────

    @staticmethod
    def _extract_paragraphs(doc: Document) -> list[dict[str, str]]:
        result = []
        for para in doc.paragraphs:
            text = para.text.strip()
            if not text:
                continue
            style_name = para.style.name if para.style else "Normal"
            level = 0
            if "Heading" in style_name:
                try:
                    level = int(style_name.split()[-1])
                except (ValueError, IndexError):
                    level = 1
            result.append({"style": style_name, "text": text, "level": str(level)})
        return result

    @staticmethod
    def _extract_tables(doc: Document) -> list[list[list[str]]]:
        tables = []
        for table in doc.tables:
            rows = []
            for row in table.rows:
                cells = []
                for cell in row.cells:
                    # Merge paragraphs within a cell into one string
                    cell_text = " ".join(p.text.strip() for p in cell.paragraphs if p.text.strip())
                    cells.append(cell_text)
                rows.append(cells)
            tables.append(rows)
        return tables

    @staticmethod
    def _build_full_text(
        paragraphs: list[dict[str, str]],
        tables: list[list[list[str]]],
    ) -> str:
        """Build a plain-text rendering suitable for LLM ingestion."""
        lines: list[str] = []

        # Walk document body in document order is complex with mixed paragraphs/tables.
        # For simplicity emit all paragraphs first, then all tables.
        # TODO: Preserve interleaved order via XML iteration if needed.

        for para in paragraphs:
            level = int(para.get("level", 0))
            prefix = "#" * level + " " if level > 0 else ""
            lines.append(f"{prefix}{para['text']}")

        if tables:
            lines.append("\n--- TABLES ---")
            for t_idx, table in enumerate(tables):
                lines.append(f"\n[Table {t_idx + 1}]")
                for row in table:
                    lines.append(" | ".join(row))

        return "\n".join(lines)
