backend/requirements.txt
# ============================================================
# LIMS OCR Backend - Python Dependencies
# Updated for Python 3.14 compatibility (unpinned versions)
# ============================================================

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9

# Document Processing
pdfplumber>=0.11.3
pdf2image>=1.17.0
pytesseract>=0.3.13
python-docx>=1.1.2
Pillow>=10.4.0

# AI / LLM
langchain>=0.3.3
langchain-openai>=0.2.3
langchain-anthropic>=0.2.3
langchain-community>=0.3.3
langchain-text-splitters>=0.3.0
openai>=1.51.0
anthropic>=0.36.0
tiktoken>=0.8.0

# Data / Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0
pandas>=2.2.3
numpy>=2.1.2

# Excel Generation
openpyxl>=3.1.5

# Database (Job History)
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# File Handling
aiofiles>=24.1.0

# Logging & Monitoring
structlog>=24.4.0

# Configuration
python-dotenv>=1.0.1
pyyaml>=6.0.2

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
httpx>=0.27.2

# Utilities
tenacity>=9.0.0
